mmb_pf/requirements.txt
asgiref
attrs
certifi
chardet
Django
django-ckeditor
django-filter
django-js-asset
django-jsoneditor
django-private-storage
djangorestframework
gevent
greenlet
grequests
idna
packaging
Pillow
psycopg2
pyparsing
python-magic
python-memcached
pytz
requests
six
sqlparse
typing-extensions
urllib3
uWSGI
zope.event
zope.interface
pytz
types-pytz